+++ Android/src/org/droidplanner/activities/EditorActivity.java
@@ -352,8 +352,4 @@ public class EditorActivity extends SuperUI implements OnPathFinishedListener,
 		updateMapPadding();
 	}
 
-	@Override
-	public CharSequence[][] getHelpItems() {
-		return new CharSequence[][] { {}, {} };
-	}
 }