+++ Android/src/org/droidplanner/activities/FlightActivity.java
@@ -1,7 +1,6 @@
 package org.droidplanner.activities;
 
 import org.droidplanner.R;
-import org.droidplanner.activities.helpers.SuperUI;
 import org.droidplanner.drone.Drone;
 import org.droidplanner.drone.DroneInterfaces.DroneEventsType;
 import org.droidplanner.drone.DroneInterfaces.OnDroneListener;
@@ -30,7 +29,7 @@ import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GooglePlayServicesUtil;
 import com.google.android.gms.maps.model.LatLng;
 
-public class FlightActivity extends SuperUI implements
+public class FlightActivity extends DrawerNavigationUI implements
 		OnMapInteractionListener, FlightActionsFragment.OnMissionControlInteraction, OnDroneListener{
 
     private static final int GOOGLE_PLAY_SERVICES_REQUEST_CODE = 101;