+++ Android/src/org/droidplanner/utils/Constants.java
@@ -44,6 +44,21 @@ public class Constants {
     public static final String PREF_UI_SCREEN = "pref_ui";
 
     /**
+     * Sets whether or not the default language for the ui should be english.
+     */
+    public static final String PREF_UI_LANGUAGE = "pref_ui_language_english";
+
+    /**
+     * By default, the system language should be used for the ui.
+     */
+    public static final boolean DEFAULT_PREF_UI_LANGUAGE = false;
+
+    /**
+     * Preference key for the drone settings' category.
+     */
+    public static final String PREF_DRONE_SETTINGS = "pref_drone_settings";
+
+    /**
      * Private constructor to prevent instantiation of this class.
      */
     private Constants() {