+++ Android/src/org/droidplanner/utils/Utils.java
@@ -1,5 +1,7 @@
 package org.droidplanner.utils;
 
+import java.util.Locale;
+
 import org.droidplanner.connection.BluetoothConnection;
 import org.droidplanner.connection.MAVLinkConnection;
 import org.droidplanner.connection.TcpConnection;
@@ -7,6 +9,9 @@ import org.droidplanner.connection.UdpConnection;
 import org.droidplanner.connection.UsbConnection;
 
 import android.content.Context;
+import android.content.res.Configuration;
+import android.content.res.Resources;
+import android.preference.PreferenceManager;
 
 /**
  * Contains application related functions.
@@ -55,5 +60,23 @@ public class Utils {
          * @since 1.2.0
          */
         public abstract MAVLinkConnection getConnection(Context context);
+    }
+
+    /**
+     * Used to update the user interface language.
+     * @param context Application context
+     */
+    public static void updateUILanguage(Context context){
+        final boolean isUiLanguageEnglish = PreferenceManager.getDefaultSharedPreferences
+                (context).getBoolean(Constants.PREF_UI_LANGUAGE,
+                Constants.DEFAULT_PREF_UI_LANGUAGE);
+
+        if(isUiLanguageEnglish){
+            Configuration config = new Configuration();
+            config.locale = Locale.ENGLISH;
+
+            final Resources res = context.getResources();
+            res.updateConfiguration(config, res.getDisplayMetrics());
+        }
     }
 }