+++ Core/src/org/droidplanner/mission/commands/MissionCMD.java
@@ -1,7 +1,12 @@
 package org.droidplanner.mission.commands;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.droidplanner.mission.Mission;
 import org.droidplanner.mission.MissionItem;
+
+import com.MAVLink.Messages.ardupilotmega.msg_mission_item;
 
 public abstract class MissionCMD extends MissionItem{
 	
@@ -11,6 +16,17 @@ public abstract class MissionCMD extends MissionItem{
 
 	public MissionCMD(MissionItem item) {
 		super(item);
+	}
+
+	@Override
+	public List<msg_mission_item> packMissionItem() {
+		List<msg_mission_item> list = new ArrayList<msg_mission_item>();
+		msg_mission_item mavMsg = new msg_mission_item();
+		list.add(mavMsg);
+		mavMsg.autocontinue = 1;
+		mavMsg.target_component = 1;
+		mavMsg.target_system = 1;
+		return list;
 	}
 	
 }