+++ Core/src/org/droidplanner/mission/commands/ReturnToHome.java
@@ -7,19 +7,21 @@ import org.droidplanner.mission.Mission;
 import org.droidplanner.mission.MissionItem;
 
 import com.MAVLink.Messages.ardupilotmega.msg_mission_item;
+import com.MAVLink.Messages.enums.MAV_CMD;
+import com.MAVLink.Messages.enums.MAV_FRAME;
 
 public class ReturnToHome extends MissionCMD {
 
+	private Altitude returnAltitude;
 
-	protected Altitude returnAltitude;
-
-	public ReturnToHome(Mission mission) {
-		super(mission);
+	public ReturnToHome(MissionItem item) {
+		super(item);
 		returnAltitude = new Altitude(0);
 	}
-	
-	public ReturnToHome(MissionItem item) {
-		this(item.getMission());
+
+	public ReturnToHome(msg_mission_item msg, Mission mission) {
+		super(mission);
+		unpackMAVMessage(msg);
 	}
 
 	public Altitude getHeight() {
@@ -32,14 +34,17 @@ public class ReturnToHome extends MissionCMD {
 
 	@Override
 	public List<msg_mission_item> packMissionItem() {
-		// TODO Auto-generated method stub
-		return null;
+		List<msg_mission_item> list = super.packMissionItem();
+		msg_mission_item mavMsg = list.get(0);
+		mavMsg.command = MAV_CMD.MAV_CMD_NAV_RETURN_TO_LAUNCH;
+		mavMsg.frame = MAV_FRAME.MAV_FRAME_GLOBAL_RELATIVE_ALT;
+		mavMsg.z = (float) returnAltitude.valueInMeters();
+		return list;
 	}
 
 	@Override
 	public void unpackMAVMessage(msg_mission_item mavMessageItem) {
-		// TODO Auto-generated method stub
-		
+		returnAltitude = new Altitude(mavMessageItem.z);
 	}
 
 }